Tasks are now user specific

+++ routes/tasksDB.js
@@ -1,19 +1,15 @@
 const express = require('express');
 const router = express.Router();
-const mongoose = require('mongoose');
 const {Task, schema} = require('../models/task.js');
 const auth = require('../middleware/auth');
 
-// Connect to MongoDB
-mongoose.connect('mongodb://127.0.0.1/taskDB')
-    .then(() => console.log('Connected to MongoDB...\n'))
-    .catch(err => console.error('Could not connect to MongoDB...', err));
 
 // Create a task
 router.post('/', auth, async (req, res) => {
     let task = {
         Title: req.body.Title,
         Task: req.body.Task,
+        Owner: req.user._id,
         AdditionalInfo: req.body.AdditionalInfo,
         Category: req.body.Category,
         Tags: req.body.Tags,
@@ -35,10 +31,14 @@ router.post('/', auth, async (req, res) => {
     }
 });
 
-// Read all tasks
+// Read all tasks by author
 router.get('/', auth, async (req, res) => {
-    const tasks = await Task.find();
-    res.send(tasks);
+    try {
+        const tasks = await Task.find({Owner: req.user._id});
+        res.send(tasks);
+    } catch (err) {
+        res.status(400).send(err.message);
+    }
 });
 
 // Read a single task
@@ -46,6 +46,9 @@ router.get('/:id', auth, async (req, res) => {
     const task = await Task.findById(req.params.id);
     if (!task) {
         return res.status(404).send('The task with the given ID was not found.');
+    }
+    if(task.Owner != req.user._id) {
+        return res.status(401).send('Access denied.');
     }
     res.send(task);
 });
@@ -55,6 +58,7 @@ router.put('/:id', auth, async (req, res) => {
     let requestedTask = {
         Title: req.body.Title,
         Task: req.body.Task,
+        Owner: req.user._id,
         AdditionalInfo: req.body.AdditionalInfo,
         Category: req.body.Category,
         Tags: req.body.Tags,
@@ -65,9 +69,15 @@ router.put('/:id', auth, async (req, res) => {
     if (error) {
         return res.status(400).send(error.details[0].message);
     } else {
-
         try {
-            let updatedTask = await Task.findOneAndUpdate({_id: req.params.id}, value, {new: true});
+            const currentTask = await Task.findById(req.params.id);
+            if (!currentTask) {
+                return res.status(404).send('The task with the given ID was not found.');
+            }
+            if (currentTask.Owner != req.user._id) {
+                return res.status(401).send('Access denied.');
+            }
+            const updatedTask = await Task.findByIdAndUpdate(req.params.id, value, {new: true});
             res.send('Task updated successfully... \n' + updatedTask);
         } catch (err) {
             res.status(500).send(err.message);
@@ -77,11 +87,15 @@ router.put('/:id', auth, async (req, res) => {
 
 // Delete a task
 router.delete('/:id', auth, async (req, res) => {
-    const task = await Task.findByIdAndRemove(req.params.id);
-    if (!task) {
+    const currentTask = await Task.findById(req.params.id);
+    if (!currentTask) {
         return res.status(404).send('The task with the given ID was not found.');
     }
-    res.send(task);
+    if (currentTask.Owner != req.user._id) {
+        return res.status(401).send('Access denied.');
+    }
+    const removedTask = await Task.findByIdAndRemove(req.params.id);
+    res.send(removedTask);
 });
 
 module.exports = router;
